Refactor into seperate files
